[x86] Fix another trivial bug in x86 flags copy lowering that has been there for a long time.
The boolean tracking whether we saw a kill of the flags was supposed to
be per-block we are scanning and instead was outside that loop and never
cleared. It requires a quite contrived test case to hit this as you have
to have multiple levels of successors and interleave them with kills.
I've included such a test case here.

This is another bug found testing SLH and extracted to its own focused
patch.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@336876 91177308-0d34-0410-b5e6-96231b3b80d8