[X86] Fix assertion in subreg extraction
This assert fires when attempting to extract a subregister from the
global PIC base register. This virtual register SD node is not in the
VRBaseMap, so we shouldn't call getVR to look it up there. If this is a
RegisterSDNode, we should be able to use the virtual register directly.

Fixes PR38385

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@339056 91177308-0d34-0410-b5e6-96231b3b80d8