[SDAG] Add versions of computeKnownBits that return a value
Having the KnownBits as an output parameter is kind of awkward to use
and a holdover from when it was two separate APInts. Instead, just
return a KnownBits object.

I'm leaving the existing interface in place for now, since updating
the callers all at once would be thousands of lines of diff.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@340594 91177308-0d34-0410-b5e6-96231b3b80d8