[X86] Mark the ISD::SETLT/SETLE condition codes as illegal for v32i16/v64i8 to match the other vector types.
I'm having a hard time finding a test case for this, but we should be consistent here. The fact that we canonicalize all zeros and all ones constants to vXi32 and all other constants to loads makes this hard to hit the easy DAG combine infinite loop we get for some of the other types.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@341859 91177308-0d34-0410-b5e6-96231b3b80d8